requirements.txt
django
channels
daphne
google-genai
pillow
python-dotenv
whitenoise